fix: Resolve type mismatch and warnings for clean build
Critical fixes:
1. HealthResponse.uptime: u64 → u32 (matches std::process::id())
2. agent.rs: let mut → let _ (remove unused mut warning)

This ensures clean compilation on Railway with zero errors.

+++ firewallx/src/api/mod.rs
@@ -32,7 +32,7 @@ pub struct DashboardStats {
 #[derive(Serialize)]
 pub struct HealthResponse {
     pub status: String,
-    pub uptime: u64,
+    pub uptime: u32,
 }
 
 async fn health_check() -> Json<HealthResponse> {
